// ts/features/itwallet/presentation/remote/analytics/enum.ts
export enum ITW_REMOTE_SCREENVIEW_EVENTS {
  ITW_REMOTE_RP_NOT_TRUSTED_BOTTOMSHEET = "ITW_REMOTE_RP_NOT_TRUSTED_BOTTOMSHEET",
  ITW_REMOTE_DATA_SHARE = "ITW_REMOTE_DATA_SHARE",
  ITW_REMOTE_INVALID_AUTH_RESPONSE_BOTTOMSHEET = "ITW_REMOTE_INVALID_AUTH_RESPONSE_BOTTOMSHEET",
  ITW_REMOTE_UX_SUCCESS = "ITW_REMOTE_UX_SUCCESS"
}

export enum ITW_REMOTE_ACTIONS_EVENTS {
  ITW_REMOTE_UX_CONVERSION = "ITW_REMOTE_UX_CONVERSION"
}

export enum ITW_REMOTE_ERRORS_EVENTS {
  ITW_REMOTE_UNEXPECTED_FAILURE = "ITW_REMOTE_UNEXPECTED_FAILURE",
  ITW_REMOTE_IDENTITY_NEEDS_VERIFICATION = "ITW_REMOTE_IDENTITY_NEEDS_VERIFICATION",
  ITW_REMOTE_RP_NOT_TRUSTED = "ITW_REMOTE_RP_NOT_TRUSTED",
  ITW_REMOTE_REQUEST_OBJECT_FAILURE = "ITW_REMOTE_REQUEST_OBJECT_FAILURE",
  ITW_REMOTE_MANDATORY_CREDENTIAL_MISSING = "ITW_REMOTE_MANDATORY_CREDENTIAL_MISSING",
  ITW_REMOTE_MANDATORY_CREDENTIAL_NOT_VALID = "ITW_REMOTE_MANDATORY_CREDENTIAL_NOT_VALID",
  ITW_REMOTE_INVALID_AUTH_RESPONSE = "ITW_REMOTE_INVALID_AUTH_RESPONSE",
  ITW_REMOTE_GENERIC_FAILURE = "ITW_REMOTE_GENERIC_FAILURE",
  ITW_REMOTE_LINK_FAILURE = "ITW_REMOTE_LINK_FAILURE",
  ITW_UPGRADE_L3_MANDATORY = "ITW_UPGRADE_L3_MANDATORY"
}

export enum ITW_REMOTE_TECH_EVENTS {
  ITW_REMOTE_START = "ITW_REMOTE_START"
}